+++ khorosjx/errors/handlers.py
@@ -57,3 +57,25 @@ def check_api_response(response, request_type='get', ignore_exceptions=False):
             __raise_exception_for_status_code(status_code, request_type, result_msg)
         successful_response = False
     return successful_response
+
+
+def check_json_for_error(json_data):
+    """This function checks to see if JSON from an API response contains an error.
+
+    :param json_data: The API response data in JSON format
+    :type json_data: dict
+    :returns: None
+    :raises: SpaceNotFoundError, GETRequestError
+    """
+    try:
+        error_status_code = json_data['error']['status']
+        # Check if the error was because the space could not be found
+        if error_status_code == 404:
+            raise errors.exceptions.SpaceNotFoundError
+        else:
+            exception_msg = f"The API request failed with a {error_status_code} status code and the " + \
+                            f"following error: {json_data['error']['message']}"
+            raise errors.exceptions.GETRequestError(exception_msg)
+    except KeyError:
+        pass
+    return